simplified caching by removing abs time expiration

+++ Cache.cpp
@@ -6,7 +6,7 @@ Cache::Cache(int size) {
 }
 
 // MTH simplified to only LRUCache
-CachePtr Cache::getCache(CacheType c, int size) {
+CachePtr Cache::getCache(CacheType c, size_t size) {
 	return make_shared<LRUCache>(size);
 }
